Rename `TypicalStudents` to `TypicalAddressBook`
