+++ src/test/java/seedu/address/logic/commands/FindCommandTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static seedu.address.logic.Messages.MESSAGE_STUDENTS_LISTED_OVERVIEW;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.testutil.TypicalStudents.getTypicalAddressBookInclTutorials;
+import static seedu.address.testutil.TypicalAddressBook.getTypicalAddressBookInclTutorials;
 
 import java.util.Arrays;
 import java.util.Collections;
