+++ src/test/java/seedu/address/storage/json/JsonAdaptedStudentTest.java
@@ -3,7 +3,7 @@ package seedu.address.storage.json;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.storage.json.JsonAdaptedStudent.MISSING_FIELD_MESSAGE_FORMAT;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalStudents.BENSON;
+import static seedu.address.testutil.TypicalAddressBook.BENSON;
 
 import java.util.ArrayList;
 import java.util.List;
